+++ backend/src/main/java/com/mealchemy/vault/repository/VaultFolderRecipeRepository.java
@@ -10,8 +10,8 @@ import java.util.List;
 /* Import classes */
 import com.mealchemy.vault.model.VaultFolderRecipe;
 
-public inteface VaultFolderRecipeRepository extends JpaRepository<VaultFolderRecipe, Long>
+public inteface VaultFolderRecipeRepository extends JpaRepository<VaultFolderRecipe, int>
 {
-    List<VaultFolderRecipe> findByFolderId(Long folderId);
-    List<VaultFolderRecipe> findByRecipeId(Long RecipeId);
+    List<VaultFolderRecipe> findByFolderId(int folderId);
+    List<VaultFolderRecipe> findByRecipeId(int RecipeId);
 }
